feat: Support selecting a platform to copy (for Actions)

+++ copy-files.sh
@@ -5,7 +5,8 @@ script_dir=$(dirname "${script_path}")
 destination=${1:-'..'} # read first arg, default to '..' (previous dir)
 
 confirm() {
-  read -rp "$1 (y/N): " yn
+  local prompt="$1"
+  read -rp "${prompt} (y/N): " yn
 
   yn=${yn,,}
   case "$yn" in
@@ -16,6 +17,26 @@ confirm() {
       return 1
       ;;
   esac
+}
+
+question() {
+  local prompt="$1"
+  local values="$2"
+  local default="$3"
+  local regex
+
+  while true; do
+    read -rp "${prompt} (values: ${values})? [default: ${default}]: " answer
+      answer=${answer:-$default}
+
+      regex="^($(echo "$values" | sed 's/, */|/g'))$"
+
+      if [[ "${answer}" =~ ${regex} ]]; then
+        echo "${answer}" && return 0
+      else
+        >&2 echo ":: Invalid input. Please choose one of the allowed values."
+      fi
+  done
 }
 
 copy_file() {
@@ -71,15 +92,17 @@ copy_fastlane() {
 
 # Copy GitHub actions (with confirmation)
 copy_github_actions() {
+  local platform="$1"
+
   android_action='.github/actions/fastlane-android'
-  if confirm ":: Copy GitHub actions for Android (${android_action})?"; then
+  if [[ "${platform}" =~ (android|all) ]] && confirm ":: Copy GitHub actions for Android (${android_action})?"; then
     [[ -d "${destination:?}/${android_action}" ]] && rm -rf "${destination:?}/${android_action}"
     mkdir -p "${destination:?}/.github/actions"
     cp -r "${android_action}" "${destination:?}/${android_action}"
   fi
 
   ios_action='.github/actions/fastlane-ios'
-  if confirm ":: Copy GitHub actions for iOS (${ios_action})?"; then
+  if [[ "${platform}" =~ (ios|all) ]] && confirm ":: Copy GitHub actions for iOS (${ios_action})?"; then
     [[ -d "${destination:?}/${ios_action}" ]] && rm -rf "${destination:?}/${ios_action}"
     mkdir -p "${destination:?}/.github/actions"
     cp -r "${ios_action}" "${destination:?}/${ios_action}"
@@ -89,8 +112,11 @@ copy_github_actions() {
 cd "${script_dir}" || exit
 
 copy_ruby_files
+
+platform=$(question ":: Platform to copy" "android, ios, all" "all")
+
 confirm ":: Copy fastlane files?" && copy_fastlane
-confirm ":: Copy GitHub actions?" && copy_github_actions
+copy_github_actions "${platform}"
 
 cd - &> /dev/null || echo ":: Couldn't go back to previous dir" || exit
 echo ":: Finished! You can now remove this repository directory"
